@@ -1,9 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>_01.CommandPattern</RootNamespace>
+    <TargetFramework>net6.0</TargetFramework>
+    <RootNamespace>CommandPattern</RootNamespace>
   </PropertyGroup>
 
 </Project>
