@@ -1,20 +1,14 @@
 ﻿using System;
+using ValidationAttributes.Models;
 
-namespace ValidationAttributes
+namespace ValidationAttributes;
+
+public class StartUp
 {
-    public class StartUp
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            //var person = new Person
-            // (
-            //     null,
-            //     -1
-            // );
-            //
-            //bool isValidEntity = Validator.IsValid(person);
-            //
-            //Console.WriteLine(isValidEntity);
-        }
+        Person person = new(null, -1);
+
+        Console.WriteLine(Validator.IsValid(person));
     }
 }
